% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3control_service.R
\name{s3control}
\alias{s3control}
\title{AWS S3 Control}
\usage{
s3control()
}
\description{
AWS S3 Control provides access to Amazon S3 control plane operations.
}
\section{Example}{
\preformatted{s3control <- paws::s3control()
s3control$operation()
}
}

\section{Operations}{

\tabular{ll}{
\link[=s3control_delete_public_access_block]{delete_public_access_block} \tab Removes the Public Access Block configuration for an Amazon Web Services account \cr
\link[=s3control_get_public_access_block]{get_public_access_block} \tab Retrieves the Public Access Block configuration for an Amazon Web Services account \cr
\link[=s3control_put_public_access_block]{put_public_access_block} \tab Creates or modifies the Public Access Block configuration for an Amazon Web Services account
}
}

